# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Oblivion API
  version: "1.0.0"
  description: |
    REST API for **Oblivion**, a face / genetic-syndrome recognition backend.

    The system is split into 7 independent microservices. They sit behind an
    nginx **API gateway** at `http://localhost:8080`, which routes by path
    prefix to the right service — so this whole API is one origin and
    "Try it out" works without switching ports. Endpoints are grouped by the
    **tag** (= service); the direct per-service ports are noted on each tag for
    when you want to bypass the gateway.

    ## Conventions
    - Every response is wrapped in an envelope:
      `{ "status": "ok" | "error", "data": <payload> | null, "error": <string> | null }`.
    - Authenticated endpoints expect `Authorization: Bearer <access_token>`,
      a RS256 JWT issued by the auth service.
    - Roles: `doctor`, `lab`, `admin`. Lab endpoints require a `lab` token.

    ## Typical flow
    1. `POST /auth/register` then `POST /auth/login` → access token.
    2. `POST /cases` → returns a presigned S3 `upload_url`.
    3. `PUT` the image bytes to `upload_url` (MinIO/S3, no auth).
    4. `POST /cases/{id}/confirm-upload` → triggers async analysis (Kafka).
    5. Poll `GET /cases/{id}` until `status = completed`; read `diagnoses`.
servers:
  - url: http://localhost:8080
    description: API gateway (nginx) — routes by path prefix to all services
...
paths:
...
  /reference/syndromes:
    get:
      tags: [reference]
      summary: List / search syndromes
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - name: search
          in: query
          required: false
          schema: { type: string }
          description: Free-text filter on syndrome name/description.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Syndrome' }
  /reference/syndromes/{id}:
    parameters: [{ $ref: '#/components/parameters/IdPath' }]
    get:
      tags: [reference]
      summary: Get a syndrome with its linked HPO terms
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/SyndromeDetail' } }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
    Offset:
      name: offset
      in: query
      required: false
      schema: { type: integer, default: 0, minimum: 0 }
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
...
  schemas:
    Envelope:
      type: object
      description: Standard response wrapper used by every endpoint.
      required: [status]
      properties:
        status: { type: string, enum: [ok, error], example: ok }
        data: { description: Payload (shape varies by endpoint), nullable: true }
        error: { type: string, nullable: true, example: null }
...
    Syndrome:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        omim_id: { type: string, nullable: true }
        description: { type: string, nullable: true }
        prevalence: { type: string, nullable: true }
        inheritance: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    HpoTerm:
      type: object
      properties:
        id: { type: string, format: uuid }
        hpo_id: { type: string, example: "HP:0000252" }
        name: { type: string }
        definition: { type: string, nullable: true }
    SyndromeDetail:
      allOf:
        - $ref: '#/components/schemas/Syndrome'
        - type: object
          properties:
            hpo_terms:
              type: array
              items: { $ref: '#/components/schemas/HpoTerm' }
...